Created the word difficulty manager class to generate word difficulties on a scale of 0 to 100 (subject to change, cap may also not actually be at 100.)
Class has functions to calculate the difficulty using 4 components, word length, the frequency of each letter in the word, the number of consecutive vowel sequences (>= 2 vowels), and the number of consecutive consonant sequences (>=3 consonants). It then returns the difficulty and prints it to the console.

The class is instantiated in the GameManager and difficulties are calculated upon set up once the list of words is passed (difficulty is calculated all at once with the word list passed).